% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_ureasem.R
\name{data_api_ureasem}
\alias{data_api_ureasem}
\title{Helper function to call API and return data frame SPECIFICALLY FOR UREA SEM DATA}
\usage{
data_api_ureasem(key, mute_onSuccess = TRUE)
}
\arguments{
\item{key}{API key}

\item{mute_onSuccess}{Logical to suppress messages}

\item{url}{API endpoint}
}
\value{
Data frame of parsed API result
}
\description{
Helper function to call API and return data frame SPECIFICALLY FOR UREA SEM DATA
}
