% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{.set_registry}
\alias{.set_registry}
\title{Assign a new value to .api_registry in the package namespace}
\usage{
.set_registry(value)
}
\description{
Assign a new value to .api_registry in the package namespace
}
\keyword{internal}
