% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_api_url.R
\name{list_api_endpoints}
\alias{list_api_endpoints}
\title{List all dataset names in the registry}
\usage{
list_api_endpoints(type = "dataset")
}
\value{
A character vector of dataset names (may include \code{NA} entries).
}
\description{
List all dataset names in the registry
}
