% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_api_url.R
\name{add_api_url}
\alias{add_api_url}
\title{Add or override an API endpoint}
\usage{
add_api_url(name, url)
}
\arguments{
\item{name}{A unique name for the API (e.g. "my_custom_data")}

\item{url}{The API endpoint URL}
}
\value{
NULL (modifies internal registry)
}
\description{
Add or override an API endpoint
}
