% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api.R
\name{data_api}
\alias{data_api}
\title{Helper function to call API and return data frame}
\usage{
data_api(url, key, mute_onSuccess = TRUE)
}
\arguments{
\item{url}{API endpoint}

\item{key}{API key}

\item{mute_onSuccess}{Logical to suppress messages}
}
\value{
Data frame of parsed API result
}
\description{
Helper function to call API and return data frame
}
