% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alldata.R
\name{get_all_api_data}
\alias{get_all_api_data}
\title{Get all built-in API data}
\usage{
get_all_api_data(
  key = Sys.getenv("lakeAPIkey"),
  list.names = names(.api_url_registry)
)
}
\arguments{
\item{key}{API key. Default reads from the lakeAPIkey environment variable.}

\item{list.names}{list of names of dataframe to get from API. run \code{list_api_endpoints()} to list all possible API}
}
\value{
A named list of data frames.
}
\description{
This function retrieves data from all registered default API endpoints.
You can extend this by adding more with \code{add_api_url()}.
}
