% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{get_api_data}
\alias{get_api_data}
\title{Get API data by name}
\usage{
get_api_data(name, key = Sys.getenv("lakeAPIkey"), type = "small")
}
\arguments{
\item{name}{One of the known API names like "wholesale_prices"}

\item{key}{API key. Default reads from the lakeAPIkey environment variable.}
}
\value{
Data frame from API
}
\description{
Get API data by name
}
