% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_api_url.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the API URL by name}
\usage{
get_api_url(name)
}
\arguments{
\item{name}{The name of the API (e.g. "wholesale_prices")}
}
\value{
The API URL
}
\description{
Get the API URL by name
}
